refactor(blueprint): extract BlueprintLoopService from route
Move create/start orchestration (agent resolution, first-prompt delivery,
session binding, start failure roll-back) out of the HTTP route into
blueprint/service.ts so orchestrators beyond the route can reuse it.
Add optional orchestration field to BlueprintLoop for external ownership.
Route behaviour unchanged.
