+++ packages/synergy/src/blueprint/index.ts
@@ -3,3 +3,4 @@ export type { Info } from "./types"
 export { LoopEvent } from "./event"
 export { LoopError } from "./error"
 export { BlueprintLoopStore } from "./loop-store"
+export { BlueprintLoopService } from "./service"
